check for fragmented molecules

+++ flowmol/analysis/molecule_builder.py
@@ -117,7 +117,7 @@ class SampledMolecule:
     def compute_valencies(self):
         """Compute the valencies of every atom in the molecule. Returns a tensor of shape (num_atoms,)."""
         adj = torch.zeros((self.num_atoms, self.num_atoms))
-        adjusted_bond_types = self.bond_types.clone()
+        adjusted_bond_types = self.bond_types.clone().float()
         adjusted_bond_types[adjusted_bond_types == 4] = 1.5
         adjusted_bond_types = adjusted_bond_types.float()
         adj[self.bond_src_idxs, self.bond_dst_idxs] = adjusted_bond_types
